requirements.txt
Django~=5.1.3
python-dotenv>=1.0.1
requests>=2.32.3
cryptography>=44.0.0
redis>=5.2.0
PyJWT>=2.10.1